+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/odata-cli
 
+## 0.17.18
+
+### Patch Changes
+
+-   Updated dependencies [cdeb18b]
+    -   @sap-ux/axios-extension@1.24.0
+
+## 0.17.17
+
+### Patch Changes
+
+-   @sap-ux/axios-extension@1.23.1
+
 ## 0.17.16
 
 ### Patch Changes