+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.6.11
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.8
+
+## 0.6.10
+
+### Patch Changes
+
+-   @sap-ux/project-input-validator@0.6.29
+-   @sap-ux/inquirer-common@0.8.7
+
+## 0.6.9
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.6
+
+## 0.6.8
+
+### Patch Changes
+
+-   @sap-ux/project-input-validator@0.6.28
+-   @sap-ux/inquirer-common@0.8.5
+
+## 0.6.7
+
+### Patch Changes
+
+-   Updated dependencies [06bc541]
+    -   @sap-ux/inquirer-common@0.8.4
+
+## 0.6.6
+
+### Patch Changes
+
+-   @sap-ux/project-input-validator@0.6.27
+-   @sap-ux/inquirer-common@0.8.3
+
+## 0.6.5
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.2
+
+## 0.6.4
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.8.1
+
+## 0.6.3
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+
 ## 0.6.2
 
 ### Patch Changes